requirements.txt
<<<<<<< HEAD
Flask
gunicorn
psycopg[binary]
=======
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Talisman==1.1.0
Flask-Limiter==4.1.1
gunicorn==25.2.0
psycopg2-binary
reportlab==4.1.0
openpyxl==3.1.2
sqlalchemy-libsql==0.2.0
>>>>>>> 8e4f3fd (Refactor to blueprint architecture, Turso integration, production hardening)
